Retain passwords read from passwdfd in memory if redial is enabled.
Keeping passwords in memory might be a security risk, but let's leave that
choice in the hands of the user.  If they configure xl2tpd to redial then
they should expect that their password will be retained in memory.  If they
don't, then wipe it away once we no longer need it.

Without this, pppd will fail to connect on a redial with an otherwise
mysterious authentication failure.

Conflicts:
	CHANGES