Fix an uninitialize memory access.
new_tunnel() initializes its 'self' field by calling new_call() and passing
the partly-initialized struct tunnel as an argument.  new_call() checks to
see if that 'self' field is NULL.  Let's avoid any ambiguity by using calloc
to initialize the new tunnel's contents to zero before we start populating
it.  This also lets us remove a bunch of explicit initialization and
ensures that all of the 'peer' field is initialized regardless of packing.